% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krbg.R
\docType{data}
\name{krbg}
\alias{krbg}
\title{2019 Korean Budget Data}
\format{A "tibble" ("tidy" data frame) with 1,590 rows and 15 variables:}
\source{
http://newslabit.hankyung.com/article/201812187348G
}
\usage{
krbg
}
\description{
2019 Korean Budget Data
}
\examples{
  krbg
}
\keyword{datasets}
