% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krbudget2019-package.R
\docType{package}
\name{krbudget2019-package}
\alias{krbudget2019}
\alias{krbudget2019-package}
\title{krbudget2019: Provide 2019 Korean Budget Data}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Chanyub Park \email{mrchypark@gmail.com} (0000-0001-6474-2570)

}
\keyword{internal}
